[hotfix] Split the final checkpoint related tests to a separate test class.
This is necessary since the StreamTaskTest is too big and has exceed the
limitation of file size configured in checkstyle plugin.
